Changed data types to clear warnings at /W4 (mex level) on MSVC. Also cleared the MD5_ functions to use standard memset/memcpy. While at it, removed the type defines and replaces the proper instances.
Clearing warnings is fun, because you find bugs like the one on TextFile.cpp CTextFile::WriteLine. Notice how wxEmptyString has a implicit conversion to wxChar* that will always be bool "true", meaning we return that we wrote the line correctly when it failed because the function returns bool, not wxString.

+++ src/libs/common/StringFunctions.cpp
@@ -78,7 +78,7 @@ wxString TruncateFilename(const CPath& filename, size_t length, bool isFilePath)
 			path.Clear();
 		} else {
 			// Minus 6 for "[...]" + separator
-			int pathlen = length - file.Length() - 6;
+			int pathlen = (int)(length - file.Length() - 6);
 			
 			if ( pathlen > 0 ) {
 				path = wxT("[...]") + path.Right( pathlen );
